alx_travel_app/alx_travel_app/requirement.txt
# Django and related packages
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Celery and message broker
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Environment variables
python-dotenv>=1.0.0

# Email and other utilities
Pillow>=10.0.0